Moved Cacti & Munin functionality to plugins.
modified:
platform_is_ok: remove curl check (only used by Cacti/Munin)

migrated:
getCactiGraphsForObject
getCactiServers
getMuninGraphsForObject
getMuninServers
getMuninNameAndDomain
renderCactiConfig
renderCactiServersEditor
renderMuninConfig
renderMuninServersEditor
renderObjectCactiGraphs
renderObjectMuninGraphs
proxyCactiRequest
proxyMuninRequest
triggerCactiGraphs
triggerMuninGraphs
