+++ tests/bootstrap.php
@@ -3,6 +3,7 @@ global $pdo_dsn, $db_username, $db_password;
 global $remote_username, $SQLSchema, $configCache, $script_mode;
 
 $script_mode = TRUE;
+$test_mode = TRUE;
 require_once '../wwwroot/inc/ophandlers.php';
 require_once '../wwwroot/inc/init.php';
 require_once '../wwwroot/inc/interface.php';
