+++ wwwroot/inc/database.php
@@ -1405,10 +1405,6 @@ function commitResetObject ($object_id = 0)
 	commitUpdateAttrValue ($object_id, 3, "");
 	// log history
 	recordObjectHistory ($object_id);
-	# Cacti graphs
-	usePreparedDeleteBlade ('CactiGraph', array ('object_id' => $object_id));
-	# Munin graphs
-	usePreparedDeleteBlade ('MuninGraph', array ('object_id' => $object_id));
 	# Do an additional reset if needed
 	callHook ('resetObject_hook', $object_id);
 }
@@ -5813,26 +5809,6 @@ function getConfiguredQuickLinks()
 	return $ret;
 }
 
-function getCactiGraphsForObject ($object_id)
-{
-	$result = usePreparedSelectBlade
-	(
-		'SELECT server_id, graph_id, caption FROM CactiGraph WHERE object_id = ? ORDER BY server_id, graph_id',
-		array ($object_id)
-	);
-	return reindexById ($result->fetchAll (PDO::FETCH_ASSOC), 'graph_id');
-}
-
-function getMuninGraphsForObject ($object_id)
-{
-	$result = usePreparedSelectBlade
-	(
-		'SELECT server_id, graph, caption FROM MuninGraph WHERE object_id = ? ORDER BY server_id, graph',
-		array ($object_id)
-	);
-	return reindexById ($result->fetchAll (PDO::FETCH_ASSOC), 'graph');
-}
-
 function touchVLANSwitch ($switch_id)
 {
 	return usePreparedExecuteBlade
@@ -5883,26 +5859,6 @@ function scanAttrRelativeDays ($attr_id, $not_before_days, $not_after_days)
 		array ($attr_id, $not_before_days, $not_after_days)
 	);
 	return $result->fetchAll (PDO::FETCH_ASSOC);
-}
-
-function getCactiServers()
-{
-	$result = usePreparedSelectBlade
-	(
-		'SELECT id, base_url, username, password, COUNT(graph_id) AS num_graphs ' .
-		'FROM CactiServer AS CS LEFT JOIN CactiGraph AS CG ON CS.id = CG.server_id GROUP BY id'
-	);
-	return reindexById ($result->fetchAll (PDO::FETCH_ASSOC));
-}
-
-function getMuninServers()
-{
-	$result = usePreparedSelectBlade
-	(
-		'SELECT id, base_url, COUNT(MG.object_id) AS num_graphs ' .
-		'FROM MuninServer AS MS LEFT JOIN MuninGraph AS MG ON MS.id = MG.server_id GROUP BY id'
-	);
-	return reindexById ($result->fetchAll (PDO::FETCH_ASSOC));
 }
 
 function isTransactionActive()
