+++ wwwroot/inc/interface-config.php
@@ -945,122 +945,6 @@ function renderMyQuickLinks ()
 	finishPortlet();
 }
 
-function renderCactiConfig()
-{
-	$servers = getCactiServers();
-	startPortlet ('Cacti servers (' . count ($servers) . ')');
-	echo '<table cellspacing=0 cellpadding=5 align=center class=widetable>';
-	echo '<tr>' .
-		'<th>base URL</th>' .
-		'<th>username</th>' .
-		'<th>graph(s)</th>' .
-		'</tr>';
-	foreach ($servers as $server)
-		echo '<tr align=left valign=top>' .
-			'<td>' . stringForTD ($server['base_url']) . '</td>' .
-			'<td>' . stringForTD ($server['username']) . '</td>' .
-			"<td class=tdright>${server['num_graphs']}</td>" .
-			'</tr>';
-	echo '</table>';
-	finishPortlet();
-}
-
-function renderCactiServersEditor()
-{
-	function printNewItemTR ()
-	{
-		printOpFormIntro ('add');
-		echo '<tr>' .
-			'<td>' . getImageHREF ('create', 'add a new server', TRUE) . '</td>' .
-			'<td><input type=text size=48 name=base_url></td>' .
-			'<td><input type=text size=24 name=username></td>' .
-			'<td><input type=password size=24 name=password></td>' .
-			'<td>&nbsp;</td>' .
-			'<td>' . getImageHREF ('create', 'add a new server', TRUE) . '</td>' .
-			'</tr></form>';
-	}
-	echo '<table cellspacing=0 cellpadding=5 align=center class=widetable>';
-	echo '<tr>' .
-		'<th>&nbsp;</th>' .
-		'<th>base URL</th>' .
-		'<th>username</th>' .
-		'<th>password</th>' .
-		'<th>graph(s)</th>' .
-		'<th>&nbsp;</th>' .
-		'</tr>';
-	if (getConfigVar ('ADDNEW_AT_TOP') == 'yes')
-		printNewItemTR();
-	foreach (getCactiServers() as $server)
-	{
-		printOpFormIntro ('upd', array ('id' => $server['id']));
-		echo '<tr><td>';
-		if ($server['num_graphs'])
-			printImageHREF ('nodestroy', 'cannot delete, graphs exist');
-		else
-			echo getOpLink (array ('op' => 'del', 'id' => $server['id']), '', 'destroy', 'delete this server');
-		echo '</td>';
-		echo '<td><input type=text size=48 name=base_url value="' . htmlspecialchars ($server['base_url'], ENT_QUOTES, 'UTF-8') . '"></td>';
-		echo '<td><input type=text size=24 name=username value="' . htmlspecialchars ($server['username'], ENT_QUOTES, 'UTF-8') . '"></td>';
-		echo '<td><input type=password size=24 name=password value="' . htmlspecialchars ($server['password'], ENT_QUOTES, 'UTF-8') . '"></td>';
-		echo "<td class=tdright>${server['num_graphs']}</td>";
-		echo '<td>' . getImageHREF ('save', 'update this server', TRUE) . '</td>';
-		echo '</tr></form>';
-	}
-	if (getConfigVar ('ADDNEW_AT_TOP') != 'yes')
-		printNewItemTR();
-	echo '</table>';
-}
-
-function renderMuninConfig()
-{
-	$servers = getMuninServers();
-	startPortlet ('Munin servers (' . count ($servers) . ')');
-	echo '<table cellspacing=0 cellpadding=5 align=center class=widetable>';
-	echo '<tr><th>base URL</th><th>graph(s)</th></tr>';
-	foreach ($servers as $server)
-	{
-		echo '<tr align=left valign=top><td>' . stringForTD ($server['base_url']) . '</td>';
-		echo "<td class=tdright>${server['num_graphs']}</td></tr>";
-	}
-	echo '</table>';
-	finishPortlet();
-}
-
-function renderMuninServersEditor()
-{
-	function printNewItemTR()
-	{
-		printOpFormIntro ('add');
-		echo '<tr>';
-		echo '<td>' . getImageHREF ('create', 'add a new server', TRUE) . '</td>';
-		echo '<td><input type=text size=48 name=base_url></td>';
-		echo '<td>&nbsp;</td>';
-		echo '<td>' . getImageHREF ('create', 'add a new server', TRUE) . '</td>';
-		echo '</tr></form>';
-	}
-	echo '<table cellspacing=0 cellpadding=5 align=center class=widetable>';
-	echo '<tr><th>&nbsp;</th><th>base URL</th><th>graph(s)</th><th>&nbsp;</th></tr>';
-	if (getConfigVar ('ADDNEW_AT_TOP') == 'yes')
-		printNewItemTR();
-	foreach (getMuninServers() as $server)
-	{
-		printOpFormIntro ('upd', array ('id' => $server['id']));
-		echo '<tr><td>';
-		if ($server['num_graphs'])
-			printImageHREF ('nodestroy', 'cannot delete, graphs exist');
-		else
-			echo getOpLink (array ('op' => 'del', 'id' => $server['id']), '', 'destroy', 'delete this server');
-		echo '</td>';
-		echo '<td><input type=text size=48 name=base_url value="' . htmlspecialchars ($server['base_url'], ENT_QUOTES, 'UTF-8') . '"></td>';
-		echo "<td class=tdright>${server['num_graphs']}</td>";
-		echo '<td>' . getImageHREF ('save', 'update this server', TRUE) . '</td>';
-		echo '</tr></form>';
-	}
-	if (getConfigVar ('ADDNEW_AT_TOP') != 'yes')
-		printNewItemTR();
-	echo '</table>';
-}
-
 function renderPluginConfig ()
 {
 	$plugins = getPlugins ();
