+++ wwwroot/inc/interface-reports.php
@@ -758,8 +758,6 @@ function renderDataIntegrityReport ()
 		'CachedPAV-FK-vlan_id' => 'CachedPAV',
 		'CachedPNV-FK-compound' => 'CachedPNV',
 		'CachedPVM-FK-object_id' => 'CachedPVM',
-		'CactiGraph-FK-server_id' => 'CactiGraph',
-		'CactiGraph-FK-server_id' => 'CactiGraph',
 		'Dictionary-FK-chapter_id' => 'Dictionary',
 		'FileLink-File_fkey' => 'FileLink',
 		'IPv4Allocation-FK-object_id' => 'IPv4Allocation',
@@ -774,8 +772,6 @@ function renderDataIntegrityReport ()
 		'MountOperation-FK-object_id' => 'MountOperation',
 		'MountOperation-FK-old_molecule_id' => 'MountOperation',
 		'MountOperation-FK-new_molecule_id' => 'MountOperation',
-		'MuninGraph-FK-server_id' => 'MuninGraph',
-		'MuninGraph-FK-server_id' => 'MuninGraph',
 		'ObjectHistory-FK-object_id' => 'ObjectHistory',
 		'ObjectLog-FK-object_id' => 'ObjectLog',
 		'PatchCableConnectorCompat-FK-connector_id' => 'PatchCableConnectorCompat',
