+++ wwwroot/inc/navigation.php
@@ -193,8 +193,6 @@ $tab['object']['ucs'] = 'UCS';
 $tab['object']['8021qorder'] = '802.1Q order';
 $tab['object']['8021qports'] = '802.1Q ports';
 $tab['object']['8021qsync'] = '802.1Q sync';
-$tab['object']['cacti'] = 'Cacti Graphs';
-$tab['object']['munin'] = 'Munin Graphs';
 $tabhandler['object']['default'] = 'renderObject';
 $tabhandler['object']['edit'] = 'renderEditObjectForm';
 $tabhandler['object']['log'] = 'renderObjectLogEditor';
@@ -215,8 +213,6 @@ $tabhandler['object']['editvslinks'] = 'renderTripletForm';
 $tabhandler['object']['8021qorder'] = 'render8021QOrderForm';
 $tabhandler['object']['8021qports'] = 'renderObject8021QPorts';
 $tabhandler['object']['8021qsync'] = 'renderObject8021QSync';
-$tabhandler['object']['cacti'] = 'renderObjectCactiGraphs';
-$tabhandler['object']['munin'] = 'renderObjectMuninGraphs';
 $tabhandler['object']['ucs'] = 'renderEditUCSForm';
 $trigger['object']['rackspace'] = 'trigger_rackspace';
 $trigger['object']['ports'] = 'trigger_ports';
@@ -234,8 +230,6 @@ $trigger['object']['tags'] = 'trigger_tags';
 $trigger['object']['8021qorder'] = 'trigger_object_8021qorder';
 $trigger['object']['8021qports'] = 'trigger_object_8021qports';
 $trigger['object']['8021qsync'] = 'trigger_object_8021qsync';
-$trigger['object']['cacti'] = 'triggerCactiGraphs';
-$trigger['object']['munin'] = 'triggerMuninGraphs';
 $trigger['object']['ucs'] = 'trigger_ucs';
 $ophandler['object']['edit']['linkObjects'] = 'linkObjects';
 $ophandler['object']['edit']['unlinkObjects'] = 'tableHandler';
@@ -284,10 +278,6 @@ $ophandler['object']['8021qsync']['exec8021QPull'] = 'process8021QSyncRequest';
 $ophandler['object']['8021qsync']['exec8021QPush'] = 'process8021QSyncRequest';
 $ophandler['object']['8021qsync']['resolve8021QConflicts'] = 'resolve8021QConflicts';
 $ophandler['object']['8021qsync']['updPortList'] = 'update8021QPortList';
-$ophandler['object']['cacti']['add'] = 'tableHandler';
-$ophandler['object']['cacti']['del'] = 'tableHandler';
-$ophandler['object']['munin']['add'] = 'tableHandler';
-$ophandler['object']['munin']['del'] = 'tableHandler';
 $ophandler['object']['ucs']['autoPopulateUCS'] = 'autoPopulateUCS';
 $ophandler['object']['ucs']['cleanupUCS'] = 'cleanupUCS';
 $delayauth['object-8021qports-save8021QConfig'] = TRUE;
@@ -672,28 +662,6 @@ $ophandler['myaccount']['interface']['upd'] = 'saveMyPreferences';
 $ophandler['myaccount']['interface']['reset'] = 'resetMyPreference';
 $ophandler['myaccount']['qlinks']['save'] = 'saveQuickLinks';
 $interface_requires['myaccount-*'] = 'interface-config.php';
-
-$page['cacti']['title'] = 'Cacti';
-$page['cacti']['parent'] = 'config';
-$tab['cacti']['default'] = 'View';
-$tab['cacti']['servers'] = 'Manage servers';
-$tabhandler['cacti']['default'] = 'renderCactiConfig';
-$tabhandler['cacti']['servers'] = 'renderCactiServersEditor';
-$ophandler['cacti']['servers']['add'] = 'tableHandler';
-$ophandler['cacti']['servers']['del'] = 'tableHandler';
-$ophandler['cacti']['servers']['upd'] = 'tableHandler';
-$interface_requires['cacti-*'] = 'interface-config.php';
-
-$page['munin']['title'] = 'Munin';
-$page['munin']['parent'] = 'config';
-$tab['munin']['default'] = 'View';
-$tab['munin']['servers'] = 'Manage servers';
-$tabhandler['munin']['default'] = 'renderMuninConfig';
-$tabhandler['munin']['servers'] = 'renderMuninServersEditor';
-$ophandler['munin']['servers']['add'] = 'tableHandler';
-$ophandler['munin']['servers']['del'] = 'tableHandler';
-$ophandler['munin']['servers']['upd'] = 'tableHandler';
-$interface_requires['munin-*'] = 'interface-config.php';
 
 $page['cableconf']['title'] = 'Patch cables';
 $page['cableconf']['parent'] = 'config';
