+++ wwwroot/inc/upgrade.php
@@ -1949,6 +1949,29 @@ CREATE TABLE `Plugin` (
   PRIMARY KEY (`name`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci";
 
+			// if the Cacti or Munin plugins are being used, mark them as enabled, otherwise uninstall them
+			// from now on, upgrades will be handled by the plugins themselves
+			$result = $dbxlink->query ('SELECT COUNT(*) FROM CactiServer');
+			if ($result->fetch (PDO::FETCH_COLUMN, 0) > 0)
+				$query[] = "INSERT INTO Plugin VALUES ('cacti','Cacti','1.0','http://www.racktables.org/','enabled')";
+			else
+			{
+				$query[] = "DELETE FROM Config WHERE varname IN ('CACTI_LISTSRC','CACTI_RRA_ID')";
+				$query[] = "DROP TABLE `CactiGraph`";
+				$query[] = "DROP TABLE `CactiServer`";
+			}
+			unset ($result);
+			$result = $dbxlink->query ('SELECT COUNT(*) FROM MuninServer');
+			if ($result->fetch (PDO::FETCH_COLUMN, 0) > 0)
+				$query[] = "INSERT INTO Plugin VALUES ('munin','Munin','1.0','http://www.racktables.org/','enabled')";
+			else
+			{
+				$query[] = "DELETE FROM Config WHERE varname = 'MUNIN_LISTSRC'";
+				$query[] = "DROP TABLE `MuninGraph`";
+				$query[] = "DROP TABLE `MuninServer`";
+			}
+			unset ($result);
+
 			$query[] = "UPDATE Config SET varvalue = '0.20.11' WHERE varname = 'DB_VERSION'";
 			break;
 		case 'dictionary':
